Conflicts:
	libraries/tuttle/src/tuttle/host/graph/ProcessGraph.cpp

+++ 3rdParty/init.py
@@ -8,6 +8,7 @@ import urllib
 import subprocess
 import tarfile
 import zipfile
+
 from shutil import move, copystat, Error, rmtree
 
 osname		= os.name.lower()
@@ -104,14 +105,16 @@ def uncompress(filename, ext, inNewDirectory, libname, folderExtracted):
 	if not os.path.exists( os.path.join( os.getcwd(), libname)):
 		os.mkdir(os.path.join( os.getcwd() ,libname))
 	
-	print ('\nuncompress : %s -> %s\n' % ( folderExtracted, libname ) )
+	print ('\nuncompress : %s -> %s\n ' % ( folderExtracted, libname ) )
 		
 	if os.path.exists( libname + '/Jamfile.v2' ) :
 		print ( 'move %s => %s' % ( libname, folderExtracted ) ) 
 		movetree( libname , folderExtracted )
-	if os.path.exists(libname) :
-		rmtree( libname )
-	os.rename( folderExtracted, libname )
+		
+	if folderExtracted != libname :	
+		if os.path.exists(libname) :
+			rmtree( libname )
+		os.rename( folderExtracted, libname )
 
 	print ('end of uncompress\n')
 
@@ -155,7 +158,7 @@ allLibs = [
 		('ctl',         'http://sourceforge.net/projects/ampasctl/files/ctl/ctl-1.4.1/ctl-1.4.1.tar.gz/download', 'ctl-1.4.1'                  , False),
 		('boost',       'http://prdownloads.sourceforge.net/boost/boost_1_47_0.tar.bz2'                         , 'boost_1_47_0'               , False), #if not windows else ('boost'   , 'http://sourceforge.net/projects/boost/files/boost/1.47.0/boost_1_47_0.zip'                   , '', False),
 		('freetype',    'http://prdownloads.sourceforge.net/freetype/freetype-2.4.3.tar.gz'                     , 'freetype-2.4.3'             , False), #if not windows else ('freetype', 'http://prdownloads.sourceforge.net/freetype/freetype-2.4.3.tar.gz'                           , '', False),
-		('libraw',      'http://www.libraw.org/data/LibRaw-0.13.1.tar.gz'                                       , 'LibRaw-0.13.1'              , False), #if not windows else ('libraw'  , 'http://www.libraw.org/data/LibRaw-0.13.5.zip'                                                , '', False),
+		('libraw',      'http://www.libraw.org/data/LibRaw-0.14.5.tar.gz'                                       , 'LibRaw-0.14.5'              , False), #if not windows else ('libraw'  , 'http://www.libraw.org/data/LibRaw-0.13.5.zip'                                                , '', False),
 		('libcaca',     'http://caca.zoy.org/files/libcaca/libcaca-0.99.beta17.tar.gz'                          , 'libcaca-0.99.beta17'        , False),
 		('lcms',        'http://downloads.sourceforge.net/project/lcms/lcms/2.1/lcms2-2.1.tar.gz'               , 'lcms2-2.1'                  , False), #if not windows else ('lcms'    , 'http://downloads.sourceforge.net/project/lcms/lcms/2.1/lcms2-2.1.zip'                        , '', False),
 		('bzlib',       'http://www.bzip.org/1.0.6/bzip2-1.0.6.tar.gz'                                          , 'bzip2-1.0.6'                , False),
@@ -169,6 +172,7 @@ allLibs = [
 		('openjpeg',    'http://openjpeg.googlecode.com/files/openjpeg_v1_4_sources_r697.tgz'                   , 'openjpeg_v1_4_sources_r697' , False),
 		('ffmpeg',      'http://ffmpeg.org/releases/ffmpeg-0.8.2.tar.bz2'                                       , 'ffmpeg-0.8.2'               , False),
 		('openimageio', 'https://github.com/OpenImageIO/oiio/tarball/RB-0.10/OpenImageIO-oiio-754a61b.tar.gz'   , 'OpenImageIO-oiio-ad1950d'   , False),
+		('opencolorio', 'http://github.com/imageworks/OpenColorIO/tarball/v1.0.4/imageworks-OpenColorIO-v1.0.4-0-gc0a9a92.tar.gz'   , 'imageworks-OpenColorIO-12fb440'   , False),
 	]
 
 
@@ -190,6 +194,24 @@ def insertInFile( filename, linesIndexes, textToAppend ):
 	fileTmp.close()
 	os.remove(filename)
 	os.rename('tmp.txt', filename)
+
+def commentInFile( filename, atLine ): 
+	file = open( filename , 'r' )
+	fileTmp = open( 'tmp.txt' , 'w' )
+	lines = file.readlines()
+
+	nlist = list( lines )
+
+	commentedLine = nlist[atLine]
+	commentedLine = "//" + commentedLine
+	nlist[atLine]=commentedLine
+	fileTmp.writelines( nlist )
+
+	file.close()
+	fileTmp.close()
+	os.remove(filename)
+	os.rename('tmp.txt', filename)
+	
 
 def makeModificationIfNecessaryInFile( filename, headerToInsert, atLine ):
 	print ( '%s' % filename )
@@ -228,6 +250,18 @@ filename = "ctl/IlmCtlSimd/CtlSimdReg.h"
 headerToInsert = "#include <cstring>"
 atLine = 55
 makeModificationIfNecessaryInFile(filename, headerToInsert, atLine );
+
+#OpenColorIO third parties extraction
+
+download_dir = "opencolorio/ext"
+
+uncompress("tinyxml_2_6_1.tar.gz", "tar.gz", False, "tinyxml", "tinyxml")
+uncompress("yaml-cpp-r482.tar.gz", "tar.gz", False, "yaml", "yaml-cpp-r482")
+
+#patch yaml
+filename = "yaml/src/emitter.cpp" 
+atLine = 663
+commentInFile(filename, atLine );
 
 print ( 'End of initialisation.' )
